refactor: platform_code/ui/core: Extract platform-dependent Screen.
