+++ zulipterminal/platform_code.py
@@ -3,9 +3,11 @@ Detection of supported platforms & platform-specific functions
 """
 
 import platform
+import re
 import subprocess
-from typing import Tuple
+from typing import Any, Tuple
 
+import urwid
 from typing_extensions import Literal
 
 
@@ -70,6 +72,20 @@ MOUSE_SELECTION_KEY = "Fn + Alt" if PLATFORM == "MacOS" else "Shift"
 def detected_platform() -> str:
     print(f"{PLATFORM} ({PLATFORM_DETAIL})")
     return PLATFORM
+
+
+class Screen(urwid.raw_display.Screen):
+    def write(self, data: Any) -> None:
+        if PLATFORM == "WSL":
+            # replace urwid's SI/SO, which produce artifacts under WSL.
+            # https://github.com/urwid/urwid/issues/264#issuecomment-358633735
+            # Above link describes the change.
+            data = re.sub("[\x0e\x0f]", "", data)
+        super().write(data)
+
+
+def generate_screen() -> Screen:
+    return Screen()
 
 
 def notify(title: str, text: str) -> str:
