+++ zulipterminal/ui.py
@@ -3,7 +3,6 @@ Defines the `View`, and controls where each component is displayed
 """
 
 import random
-import re
 import time
 from typing import Any, Dict, List, Optional
 
@@ -23,7 +22,7 @@ from zulipterminal.config.symbols import (
 )
 from zulipterminal.config.ui_sizes import LEFT_WIDTH, RIGHT_WIDTH, TAB_WIDTH
 from zulipterminal.helper import asynch
-from zulipterminal.platform_code import MOUSE_SELECTION_KEY, PLATFORM
+from zulipterminal.platform_code import MOUSE_SELECTION_KEY
 from zulipterminal.ui_tools.boxes import MessageSearchBox, WriteBox
 from zulipterminal.ui_tools.views import (
     LeftColumnView,
@@ -348,13 +347,3 @@ class View(urwid.WidgetWrap):
             self.displaying_selection_hint = False
 
         return super().mouse_event(size, event, button, col, row, focus)
-
-
-class Screen(urwid.raw_display.Screen):
-    def write(self, data: Any) -> None:
-        if PLATFORM == "WSL":
-            # replace urwid's SI/SO, which produce artifacts under WSL.
-            # https://github.com/urwid/urwid/issues/264#issuecomment-358633735
-            # Above link describes the change.
-            data = re.sub("[\x0e\x0f]", "", data)
-        super().write(data)
